trading-service/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==0.19.0
requests==2.26.0
websockets==9.1
pydantic==1.8.2
redis==3.5.3
pika==1.2.0
jaeger-client==4.7.0
prometheus-client==0.11.0
python-consul==1.1.0
kafka-python==2.0.2
pyjwt==2.1.0
python-binance==1.0.15
sqlalchemy==1.4.23
psycopg2-binary==2.9.1 